workers/image_processor/requirements.txt
Pillow==10.2.0
pika==1.3.2
boto3==1.34.34
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
